Refactor files/imports and fix some bugs
