+++ frontend/src/api/hello.ts
@@ -1,4 +1,8 @@
-const fetchHello = (): Promise<string> => fetch('/api/v1/hello')
-  .then((response) => response.text());
+type HelloMessage = {
+  message: string;
+}
+
+const fetchHello = (): Promise<HelloMessage> => fetch('/api/v1/hello')
+  .then((response) => response.json());
 
 export default fetchHello;
