+++ frontend/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createGlobalStyle } from 'styled-components';
-import App from './App';
+import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 
 const GlobalStyle = createGlobalStyle`
